refactor: js에서 ts파일로 변환
